Move VerticalGapLayout to .~/components,
to reuse it between pages.

(redo commit 3eafa2bd493c8e5a9eebf38c1eb2f675109c71d9)
