+++ js/src/components/vertical-gap-layout/index.scss
