+++ js/src/setup-mc/setup-stepper/choose-audience/form-content.js
@@ -15,7 +15,7 @@ import Subsection from '.~/wcdl/subsection';
 import RadioHelperText from '.~/wcdl/radio-helper-text';
 import StepContentFooter from '.~/components/stepper/step-content-footer';
 import SupportedCountrySelect from './supported-country-select';
-import VerticalGapLayout from '../setup-free-listings/components/vertical-gap-layout';
+import VerticalGapLayout from '.~/components/vertical-gap-layout';
 import useAutoSaveTargetAudienceEffect from './useAutoSaveTargetAudienceEffect';
 import useAutoClearShippingEffect from './useAutoClearShippingEffect';
 import './index.scss';
