+++ js/src/setup-mc/setup-stepper/setup-free-listings/pre-launch-checklist/index.js
@@ -12,7 +12,7 @@ import { CheckboxControl } from '@wordpress/components';
 import HelpPopover from '.~/components/help-popover';
 import AppDocumentationLink from '.~/components/app-documentation-link';
 import Section from '.~/wcdl/section';
-import VerticalGapLayout from '../components/vertical-gap-layout';
+import VerticalGapLayout from '.~/components/vertical-gap-layout';
 import './index.scss';
 
 const PreLaunchChecklist = ( props ) => {
