+++ js/src/setup-mc/setup-stepper/setup-free-listings/shipping-rate/shipping-rate-setup/add-rate-button/add-rate-modal/index.js
@@ -12,7 +12,7 @@ import AppModal from '.~/components/app-modal';
 import AppInputControl from '.~/components/app-input-control';
 import useStoreCurrency from '.~/hooks/useStoreCurrency';
 import { useAppDispatch } from '.~/data';
-import VerticalGapLayout from '../../../../components/vertical-gap-layout';
+import VerticalGapLayout from '.~/components/vertical-gap-layout';
 import AudienceCountrySelect from '../../../../components/audience-country-select';
 import useGetRemainingCountryCodes from './useGetRemainingCountryCodes';
