+++ js/src/setup-mc/setup-stepper/setup-free-listings/shipping-rate/shipping-rate-setup/countries-price-input/edit-rate-button/edit-rate-modal/index.js
@@ -11,7 +11,7 @@ import { Form } from '@woocommerce/components';
 import { useAppDispatch } from '.~/data';
 import AppModal from '.~/components/app-modal';
 import AppInputControl from '.~/components/app-input-control';
-import VerticalGapLayout from '../../../../../components/vertical-gap-layout';
+import VerticalGapLayout from '.~/components/vertical-gap-layout';
 import AudienceCountrySelect from '../../../../../components/audience-country-select';
 import './index.scss';
