+++ js/src/setup-mc/setup-stepper/setup-free-listings/shipping-time/shipping-time-setup/add-time-button/add-time-modal/index.js
@@ -10,7 +10,7 @@ import { Form } from '@woocommerce/components';
  */
 import AppModal from '.~/components/app-modal';
 import AppInputControl from '.~/components/app-input-control';
-import VerticalGapLayout from '../../../../components/vertical-gap-layout';
+import VerticalGapLayout from '.~/components/vertical-gap-layout';
 import AudienceCountrySelect from '../../../../components/audience-country-select';
 import { useAppDispatch } from '.~/data';
 import useGetRemainingCountryCodes from './useGetRemainingCountryCodes';
